Create Final Entry Project "Fruit Store"
This one was a long undertaking, took reviewing the module a few times for each step to be certain of my queries, but overall very satisfied with my finished project. It will be one to remember when I am able to skim through this and know what each step means and be able to recreate much quicker than my first try.